TelegramBots/DemoTelgramBot/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
aiofiles==23.2.1
    # via
    #   -r requirements.in
    #   aiogram
aiogram==3.4.1
    # via -r requirements.in
aiohttp==3.9.5
    # via
    #   aiogram
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   openai
asttokens==2.4.1
    # via stack-data
async-timeout==4.0.3
    # via
    #   aiohttp
    #   langchain
attrs==23.2.0
    # via aiohttp
certifi==2024.2.2
    # via
    #   aiogram
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
comm==0.2.2
    # via ipykernel
contourpy==1.2.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
dataclasses-json==0.6.4
    # via
    #   langchain
    #   langchain-community
debugpy==1.8.1
    # via ipykernel
decorator==5.1.1
    # via ipython
distro==1.9.0
    # via openai
exceptiongroup==1.2.1
    # via
    #   anyio
    #   ipython
executing==2.0.1
    # via stack-data
faiss-cpu==1.8.0
    # via -r requirements.in
fonttools==4.51.0
    # via matplotlib
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via openai
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==7.1.0
    # via jupyter-client
importlib-resources==6.4.0
    # via matplotlib
ipykernel==6.29.4
    # via -r requirements.in
ipython==8.18.1
    # via ipykernel
jedi==0.19.1
    # via ipython
jsonpatch==1.33
    # via
    #   langchain
    #   langchain-core
jsonpointer==2.4
    # via jsonpatch
jupyter-client==8.6.1
    # via ipykernel
jupyter-core==5.7.2
    # via
    #   ipykernel
    #   jupyter-client
kiwisolver==1.4.5
    # via matplotlib
langchain==0.1.16
    # via -r requirements.in
langchain-community==0.0.34
    # via
    #   -r requirements.in
    #   langchain
langchain-core==0.1.46
    # via
    #   langchain
    #   langchain-community
    #   langchain-openai
    #   langchain-text-splitters
langchain-openai==0.1.3
    # via -r requirements.in
langchain-text-splitters==0.0.1
    # via langchain
langsmith==0.1.51
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
magic-filter==1.0.12
    # via aiogram
marshmallow==3.21.1
    # via dataclasses-json
matplotlib==3.8.4
    # via -r requirements.in
matplotlib-inline==0.1.7
    # via
    #   ipykernel
    #   ipython
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
nest-asyncio==1.6.0
    # via ipykernel
numpy==1.26.4
    # via
    #   contourpy
    #   faiss-cpu
    #   langchain
    #   langchain-community
    #   matplotlib
openai==1.23.2
    # via
    #   -r requirements.in
    #   langchain-openai
orjson==3.10.1
    # via langsmith
packaging==23.2
    # via
    #   ipykernel
    #   langchain-core
    #   marshmallow
    #   matplotlib
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pillow==10.3.0
    # via matplotlib
platformdirs==4.2.1
    # via jupyter-core
prompt-toolkit==3.0.43
    # via ipython
psutil==5.9.8
    # via ipykernel
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pydantic==2.5.3
    # via
    #   aiogram
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
pydantic-core==2.14.6
    # via pydantic
pygments==2.17.2
    # via ipython
pyparsing==3.1.2
    # via matplotlib
pypdf==4.2.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   jupyter-client
    #   matplotlib
python-dotenv==1.0.1
    # via -r requirements.in
pyyaml==6.0.1
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
pyzmq==26.0.2
    # via
    #   ipykernel
    #   jupyter-client
regex==2024.4.16
    # via tiktoken
requests==2.31.0
    # via
    #   langchain
    #   langchain-community
    #   langsmith
    #   tiktoken
six==1.16.0
    # via
    #   asttokens
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
sqlalchemy==2.0.29
    # via
    #   langchain
    #   langchain-community
stack-data==0.6.3
    # via ipython
tenacity==8.2.3
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
tiktoken==0.6.0
    # via
    #   -r requirements.in
    #   langchain-openai
tornado==6.4
    # via
    #   ipykernel
    #   jupyter-client
tqdm==4.66.2
    # via openai
traitlets==5.14.3
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
typing-extensions==4.11.0
    # via
    #   aiogram
    #   anyio
    #   ipython
    #   openai
    #   pydantic
    #   pydantic-core
    #   pypdf
    #   sqlalchemy
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
urllib3==2.2.1
    # via requests
wcwidth==0.2.13
    # via prompt-toolkit
yarl==1.9.4
    # via aiohttp
zipp==3.18.1
    # via
    #   importlib-metadata
    #   importlib-resources
